V2.0.3.23-2
1. Code optimize
2. Optimize runtime in  OnAfterSimulationFrame
